--- src/components/containers/AnswerCardContainer.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AnswerCardContainer = () => {
-  return <div></div>;
-};
-
-export default AnswerCardContainer;
