+++ src/components/ui/AnswerFeed.jsx
@@ -1,0 +1,5 @@
+function AnswerFeed() {
+  return <div></div>;
+}
+
+export default AnswerFeed;
