+++ src/components/ui/ReactionPresenter.jsx
@@ -1,0 +1,5 @@
+function ReactionPresenter() {
+  return <div></div>;
+}
+
+export default ReactionPresenter;
